refactor: change validate sdk setup script to use env vars
